refactor(core): split away connection arg creation
